==> code/3a_DataCleaning_ResaleFlat.ipynb <==
# Import Libaries

# Import Libaries
import pandas as pd
import numpy as np

# Data Cleaning for Resale Flat data with longitude and latitude information

- The resale flat price are retrieved from the from the following website from 1990 onwards to July 2023 [[1]](https://beta.data.gov.sg/datasets/189/resources/d_8b84c4ee58e3cfc0ece0d773c8ca6abc/view)

- The dataset has been combined with longitude and latitude data extracted from onemap [[2]](https://www.onemap.gov.sg/apidocs/apidocs)

## Import Dataset

# Import in Dataset
resale_long_lat_fmt_only = pd.read_csv('../assets/resale_flat_gen/resale_long_lat_fmt_only.csv')


#
resale_long_lat_fmt_only.info()

## Check if there are null values in the dataset

# Check if there are missing values for dataset
resale_long_lat_fmt_only.isnull().sum()

Missing values is seen for the column "remaining_lease".

# View the top 5 rows 
resale_long_lat_fmt_only.head(5)

# Filter out the unique months where remaining lease is NaN
resale_month_NaN = resale_long_lat_fmt_only[resale_long_lat_fmt_only["remaining_lease"].isnull()]["month"]
resale_month_NaN.unique()

Observed that data missing the resale price from 1990 to December 2014.

# Retrieve 5 rows where remaining lease is not null to view how it is calculated.
resale_long_lat_fmt_only[resale_long_lat_fmt_only["remaining_lease"].isnull() == False].head(5)

Observed that the remaining lease requires the current year of row to derive the value.
Next: Column "Year" and "Month" will be added for purposes such as EDA and derivation of values such as "remaining_lease"

## Create Columns for Data Frame

# add month column

# create column with month column formatted to date time format
resale_long_lat_fmt_only['month_fmt'] = pd.to_datetime(resale_long_lat_fmt_only['month'])

# drop "month" coumn
resale_long_lat_fmt_only.drop(['month'], axis=1, inplace = True)

# create month column
resale_long_lat_fmt_only["month"] = resale_long_lat_fmt_only['month_fmt'].dt.month

# create year column
resale_long_lat_fmt_only["year"] = resale_long_lat_fmt_only['month_fmt'].dt.year

# check info and top 3 rows with newly create month and year column

# check info:
print(resale_long_lat_fmt_only.info())

# check first 3 rows:
print(resale_long_lat_fmt_only.head(3))

Observed that year and month column has been created successfully. The original column of month where data is in the format of "yyyy-mm" has been drop.


## Calculate Remaining Lease

Remaining lease is calculated from the difference between 99 and the current age of the house based on current year

resale_long_lat_fmt_only["remaining_lease"] = 99 - (resale_long_lat_fmt_only["year"] 
                                             - resale_long_lat_fmt_only["lease_commence_date"])

# View dataset after calculating remaining lease value:
print(resale_long_lat_fmt_only.info())

# View first 3 rows
print(resale_long_lat_fmt_only.head(3))

Observed that remaining lease values has been calculated and there is no more null values for  the current values.

# Joined Dataset with Planning areas for cleaning

Planning area is derived from the OneMap website.
The datarows tally with with the original dataset which has been referenced to derive the information

# Import In dataset containing the Planned Areas

# Import in Dataset
planning_area_df = pd.read_csv('../assets/onemap/planning_area_query.csv')


## Check if there are null values in the dataset

# get dataset info
planning_area_df.info()

Observed that there are missing values for the column "planning_query_town".

This is expected as this column is derived from the onemap api based on given latitude and longitude information.

# Get first 10 rows of data to view relationship between "town" and "planning_query_town"
planning_area_df.head(10)

## Compare unique values of "town" and "planning_query_town" to observed if there are similarities

# Get unique "town" values
unique_town = planning_area_df["town"].unique()
unique_town

# Get unique "planning_query_town" values
unqiue_town_planning_query = planning_area_df["planning_query_town"].unique()
unqiue_town_planning_query

# Get unique_town_planning_query without NaN value
len_query= len(unqiue_town_planning_query)
unique_town_planning_query = unqiue_town_planning_query[1:36]

# Get the differences between unqiue town value and unique value of planning_area_town
differences_town = np.setdiff1d(unique_town_planning_query, unique_town)
differences_town

- Observed that there are some similarities between the "town" column and the "planning_query_town"
- There are towns found in "planning_query_town" but not "town" such as 'CHANGI', 'DOWNTOWN CORE', 'KALLANG', 'NOVENA', 'OUTRAM', 'ROCHOR','SUNGEI KADUT', 'TANGLIN', 'TUAS'.

# Get town values where "planning_query_town" value is missing
unique_town_missing = planning_area_df[planning_area_df["planning_query_town"].isnull()]["town"].unique()
unique_town_missing


# Get the differences between unqiue town value with missing value and unique value of planning_area_town
# values that are found in derived "planning_query_town" but not in "town"
differences_town_missing = np.setdiff1d(unique_town_planning_query, unique_town_missing)
differences_town_missing

# Difference that are found "town" column but not "planning_query_town"
np.setdiff1d(unique_town_missing,unique_town_planning_query)

# Observerd the unique :town" counts which has the missing "planning_town_query" values
planning_area_df[planning_area_df["planning_query_town"].isnull()]["town"].value_counts()

- Observed that for large number of the town with missing values for planning_query_town, most of the "town" columns map with the "planning_town_query".
- 'CENTRAL AREA', 'KALLANG/WHAMPOA' are columns that are found under "town" column but not in "planning_query_town".

Next:
- KALLANG/WHAMPOA will be mapped to KALLANG,
- Understand data where town is "CENTRAL AREA" to decided what to map data as for "PLANNING_QUERY_TOWN" or to drop the rows.
    - There are 966 rows out of 903393 rows at 0.1%

## Analyse rows where town is "Central Area" and "Planning_Query_Town" is null

# Get rows from planning_area_df where planning_query_town is null
planning_area_df[(planning_area_df["town"] == "CENTRAL AREA") & planning_area_df["planning_query_town"].isnull()].head()

Observed that the planning_area_df only shows latitude longitude  
Next: to retrieve unique street name from resale flat price df to better understand the flat locations.

# Get unique street_name from resale flat where town is CENTRAL AREA
resale_long_lat_fmt_only[resale_long_lat_fmt_only["town"] == "CENTRAL AREA"]['street_name'].unique()

From the following street name where town is "CENTRAL AREA", the respective town will be mapped: 

- 'JLN KUKOH' will be mapped to 'OUTRAM' as it is near Havelock road.
- 'OUTRAM PK' will be mapped to 'OUTRAM'
- 'SHORT ST' will be mapped to 'ROCHOR' as it is near Rochor station.
- 'SELEGIE RD' will be mapped to 'ROCHOR' as it is near Rochor station.
- 'UPP CROSS ST' will be mapped to 'OUTRAM'
- 'WATERLOO ST' will be mapped to 'ROCHOR'
- 'QUEEN ST' will be mapped to 'ROCHOR'
- 'BUFFALO RD' will be mapped to 'ROCHOR'
- 'ROWELL RD' will be mapped to 'ROCHOR'
- 'ROCHOR RD'will be mapped to 'ROCHOR'
- 'BAIN ST' will be mapped to 'ROCHOR'
- 'SMITH ST' will be mapped to 'OUTRAM'
- 'VEERASAMY RD' will be mapped to 'ROCHOR'
- 'NEW MKT RD' will be mapped to 'OUTRAM'
- 'TG PAGAR PLAZA' will be mapped to 'OUTRAM'
- 'KELANTAN RD' will be mapped to 'ROCHOR'
- 'SAGO LANE' will be mapped to 'OUTRAM'
- 'JLN BERSEH' will be mapped to 'ROCHOR'
- 'CHANDER RD' will be mapped to 'ROCHOR'
- 'CHIN SWEE RD' will be mapped to 'OUTRAM'
- 'KRETA AYER RD'will be mapped to 'OUTRAM'
- 'OUTRAM HILL' will be mapped to 'OUTRAM'
- 'KLANG LANE' will be mapped to 'ROCHOR'
- 'CANTONMENT RD' will be mapped to 'OUTRAM'


# Next: To find out what are the towns with the planning_query_town data missing
planning_area_df[planning_area_df["planning_query_town"].isnull()]


# Concat Resale Flat Price Data with Town Planning Query

- "planning_town_query" column will be concatenated with with resale flat price dataset.
- The concatenated dataset will have 'street_name' available to be used as a conditional requirement to fill in 'town' column

# Add column "planning_query_town" to datafrane
resale_long_lat_fmt_only_concat_town = resale_long_lat_fmt_only.copy()
resale_long_lat_fmt_only_concat_town["planning_query_town"] = planning_area_df["planning_query_town"]

# view concatenated table
resale_long_lat_fmt_only_concat_town.info()

Observed that "planning_town_query" is now added to the dataframe

## Map Planning_Query_Town to Town

# view the rows where the plannning_query_town is NaN
resale_long_lat_fmt_only_concat_town[resale_long_lat_fmt_only_concat_town["planning_query_town"].isnull()].head(5)


### Update rows where town is "KALLANG/WHAMPOA" to "KALLANG " for "planning_query_town".

# Check count where town is "KALLANG/WHAMPOA" and "planning_query_town" is NaN
resale_long_lat_fmt_only_concat_town[
                                    (resale_long_lat_fmt_only_concat_town['town'] == 'KALLANG/WHAMPOA')
                                    & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                    ].info()

It is observed that there are 3775 rows to for "planning_query_town" area to be renamed to "KALLANG"

# Rename Column to "KALLANG"
resale_long_lat_fmt_only_concat_town.loc[(resale_long_lat_fmt_only_concat_town['town']=='KALLANG/WHAMPOA')
                                         & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                        ,'planning_query_town'] = 'KALLANG'

# check that column has been updated
resale_long_lat_fmt_only_concat_town[resale_long_lat_fmt_only_concat_town['town']=='KALLANG/WHAMPOA'].head(3)

# Check if columns have been updated successfully.
# Check count where town is "KALLANG/WHAMPOA" and "planning_query_town" is NaN
resale_long_lat_fmt_only_concat_town[
                                    (resale_long_lat_fmt_only_concat_town['town'] == 'KALLANG/WHAMPOA')
                                    & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                    ].info()

Observe and validated that "KALLANG" has been updated for "planning_query_town".

## Update planning_query_town for town "CENTRAL AREA"

- "Central Area" is not an option under "planning_query_town".
- "OUTRAM" or "ROCHOR" will be assigned as a value under "planning_query_town" depending on the "street_name" for town.

# Create List of Street Name where town is "CENTRAL AREA" for planning_query_town to be mapped as "OUTRAM"
street_name_to_OUTRAM = ['JLN KUKOH', 'OUTRAM PK', 'UPP CROSS ST'
                         , 'SMITH ST', 'NEW MKT RD', 'TG PAGAR PLAZA'
                         , 'SAGO LANE', 'CHIN SWEE RD', 'KRETA AYER RD'
                         , 'OUTRAM HILL', 'CANTONMENT RD'
                        ]

# Create List of Street Name where town is "CENTRAL AREA" for planning_query_town to be mapped as "ROCHOR"
street_name_to_ROCHOR = ['SHORT ST', 'SELEGIE RD', 'WATERLOO ST'
                         , 'QUEEN ST', 'BUFFALO RD', 'ROWELL RD'
                         , 'ROCHOR RD', 'BAIN ST', 'VEERASAMY RD'
                         , 'KELANTAN RD', 'JLN BERSEH', 'CHANDER RD'
                         , 'KLANG LANE'                        
                        ] 


# Create for loop to set "planning_query_town" to OUTRAM if street_name matches list

# instantiate counter variable
counter = 0

for street_name in street_name_to_OUTRAM:
    resale_long_lat_fmt_only_concat_town.loc[(resale_long_lat_fmt_only_concat_town['town']=='CENTRAL AREA')
                                             & (resale_long_lat_fmt_only_concat_town['street_name']== street_name)
                                             & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                            ,'planning_query_town'] = 'OUTRAM'

# # Sample check that OUTRAM has been replaced base on street_name
resale_long_lat_fmt_only_concat_town[(resale_long_lat_fmt_only_concat_town['town'] == 'CENTRAL AREA')
                                    & (resale_long_lat_fmt_only_concat_town['street_name'] == 'JLN KUKOH')].head(3)

    

# Create for loop to set "planning_query_town" to ROCHOR if street_name matches list

# instantiate counter variable
counter = 0

for street_name in street_name_to_ROCHOR:
    resale_long_lat_fmt_only_concat_town.loc[(resale_long_lat_fmt_only_concat_town['town']=='CENTRAL AREA')
                                             & (resale_long_lat_fmt_only_concat_town['street_name']== street_name)
                                             & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                            ,'planning_query_town'] = 'ROCHOR'

# # Sample check that OUTRAM has been replaced base on street_name
resale_long_lat_fmt_only_concat_town[(resale_long_lat_fmt_only_concat_town['town'] == 'CENTRAL AREA')
                                    & (resale_long_lat_fmt_only_concat_town['street_name'] == 'SHORT ST')].head(3)

# check that column has been updated
resale_long_lat_fmt_only_concat_town[resale_long_lat_fmt_only_concat_town['town']=='CENTRAL AREA'].info()

Observed that there are no more null values for planning_query_town where "town" is 'CENTRAL AREA".

### Update planning_query_town where town is the same  name

# Retrieve unique town name where planning_query_town is null.
resale_long_lat_fmt_only_concat_town[resale_long_lat_fmt_only_concat_town["planning_query_town"].isnull()]["town"].value_counts()


- Observed that the unique "town" values where "planning_query_town" is NaN has the same name as the existing "planning_query_town" as follows.

- Unique "planning_query_town":
        'ANG MO KIO', 'BEDOK', 'BISHAN', 'BUKIT BATOK', 'BUKIT MERAH',
       'BUKIT TIMAH', 'TANGLIN', 'OUTRAM', 'ROCHOR', 'DOWNTOWN CORE',
       'CHOA CHU KANG', 'CLEMENTI', 'GEYLANG', 'HOUGANG', 'JURONG EAST',
       'JURONG WEST', 'KALLANG', 'CHANGI', 'NOVENA', 'QUEENSTOWN',
       'MARINE PARADE', 'SERANGOON', 'TAMPINES', 'TOA PAYOH', 'WOODLANDS',
       'YISHUN', 'TUAS', 'SUNGEI KADUT', 'LIM CHU KANG', 'BUKIT PANJANG',
       'SEMBAWANG', 'PASIR RIS', 'SENGKANG', 'PUNGGOL'

- Next: update the remaining planning_query_town value with NaN with existing values of "town".

# Instantiate list for town where planning_query_town is NaN.
town_where_planning_query_nan = ["JURONG WEST" ,"ANG MO KIO" , "HOUGANG"
                                , "BEDOK" ,"PASIR RIS" ,"BUKIT MERAH" 
                                ,"QUEENSTOWN" ,"YISHUN" ,"GEYLANG"
                                , "JURONG EAST" ,"WOODLANDS" ,"BUKIT BATOK"
                                , "CLEMENTI" , "TAMPINES" ,"CHOA CHU KANG"
                                , "SENGKANG" ,"PUNGGOL" ,"BUKIT PANJANG"
                                , "TOA PAYOH" , "BISHAN" ,"SERANGOON"
                                , "SEMBAWANG" ,"MARINE PARADE" ,"BUKIT TIMAH"
                                , "LIM CHU KANG"]


# Loop through town list and update NaN values for planning_query_town
for town in town_where_planning_query_nan:
    resale_long_lat_fmt_only_concat_town.loc[(resale_long_lat_fmt_only_concat_town['town']== town)
                                             & (resale_long_lat_fmt_only_concat_town['planning_query_town'].isnull())
                                            ,'planning_query_town'] = resale_long_lat_fmt_only_concat_town['town']

#  Check that there is no more null values for planning_query_town.
resale_long_lat_fmt_only_concat_town.info()

# Check count of unique planning_query_town 
resale_long_lat_fmt_only_concat_town["planning_query_town"].value_counts()

- Validated that the planning_town_query column has been NaN value has been filled up.
- The total count of planning_town_query unique column sums up to 903393
- Validated that there are no null values and the data type is accurate.

## Export Cleaned Dataset

# export coordinateList to csv
resale_long_lat_fmt_only_concat_town.to_csv('../assets/data_clean/resale_w_plantown_clean.csv', encoding='utf-8', index=False)

# export as pickle file
resale_long_lat_fmt_only_concat_town.to_pickle('../assets/data_clean/resale_w_plantown_clean.pkl')

- Observed that the files are successfully exported.
