==> code/4_EDA.ipynb <==
# Import Libraries

# Import Libaries
import pandas as pd
import numpy as np

pd.set_option('display.max_columns', None)

# Functions

# Function to one hot encode selected columns  
def hot_encode_cat_col(dataframe,list_cat_col):
    for col in list_cat_col:    
        dataframe = pd.get_dummies(columns=[col],drop_first=True,data=dataframe) 
    return dataframe

# Function to loop through the column list and drops the columns
def drop_columns(dataframe,col_list):
    for col in col_list:
        dataframe.drop(columns=[col], inplace=True)
           

# Load Dataset

# Load in Dataset
combined_resale_df = pd.read_csv('../assets/data_clean/resale_combine_clean.csv')

# One Hot Encode Object Column

# view dataframe info()
combined_resale_df.info(verbose=True)

# Get columns with datatype object
string_columns = combined_resale_df.select_dtypes(include=['object']).columns.tolist()
string_columns

# View the number of unique value for each column value to decide what to encode
# combined_resale_df['town'].unique()

for string_col in string_columns:
    print(string_col,": ", len(combined_resale_df[string_col].unique()))

# view first 2 rows of data
combined_resale_df.head(2)

- Observed that block, street name and address has high unique values.
    - These columns will be dropped as location is covered by "planning_area"
- mth_format will be drop as there is already year and month column

# Drop columns
col_to_drop = ['block', 'street_name', 'address',  'month_fmt']
drop_columns(combined_resale_df,col_to_drop)

# Check that the columns has been dropped and update string_columns
string_columns = combined_resale_df.select_dtypes(include=['object']).columns.tolist()
string_columns

- Observed that the columns has been dropped
- Next: proceed with one hot encoding for remaining columns

# create a copy of dataframe for one hot encoding
combined_resale_df_copy = combined_resale_df.copy()

# One Hot Encode all String Columns
combined_resale_df_copy = hot_encode_cat_col(combined_resale_df_copy,string_columns)

# Check that there are no more string columns
combined_resale_df_copy.select_dtypes(include=['object']).columns.tolist()

- Validated that there are no more string columns where dtype is object

# Export Dataset for Modelling

# export coordinateList to csv
combined_resale_df_copy.to_csv('../assets/data_clean/resale_combine_encode_clean.csv', encoding='utf-8', index=False)

# export as pickle file
combined_resale_df_copy.to_pickle('../assets/data_clean/resale_combine_encode_clean.pkl')

- Observed that dataset has been exported

- Data will first be exported for Modelling to retrieve a baseline model for comparision. Further analysis of the dataset will be done in the subsequent notebook.
