==> code/5a_Model.ipynb <==
# Import Libraries

# Import Libaries
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split, cross_val_score
from sklearn.linear_model import LinearRegression, Lasso, Ridge
from sklearn.metrics import mean_squared_error, r2_score 
from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import GridSearchCV
from sklearn.pipeline import Pipeline

pd.set_option('display.max_columns', None)

# Load Dataset

# Load in Dataset
combined_resale_encode_df = pd.read_pickle('../assets/data_clean/resale_combine_encode_clean.pkl')

combined_resale_encode_df.info(verbose = True) 

# Plot Baseline Model

## Linear Regression Model

# Get all columns
all_column = combined_resale_encode_df.columns.tolist()

# Remove resale price from all_column 
all_column.remove('resale_price')

# Set feature list
features = all_column

# Set Feature
X = combined_resale_encode_df[features]

# Set Target Variable
y = combined_resale_encode_df['resale_price']

# Create Train Test Split 
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=123)

# Instantiate Linear Regression Model
lr = LinearRegression()

# Standardise the predictors
# Instantiating Standard Scaler
sc = StandardScaler()

# Standardise Predictors
Z_train = sc.fit_transform(X_train) # standard normal distribution
Z_test = sc.transform(X_test)

# Fit the model
lr.fit(Z_train,y_train)

# Generate predictions.
preds = lr.predict(Z_test)

y_pred_train = lr.predict(Z_train)
y_pred_test = lr.predict(Z_test)

rmse_train = mean_squared_error(y_true =y_train, y_pred=y_pred_train, squared=False)
rmse_test = mean_squared_error(y_true =y_test, y_pred=y_pred_test, squared=False)

print(rmse_train, rmse_test)

# Check the R^2 on the training and testing sets.

print(f'R^2 on testing set: {r2_score(y_train, y_pred_train)}')
print(f'R^2 on training set: {r2_score(y_test, y_pred_test)}')

- R2 Score for both Training and Testing data is 0.85. 
    - The score is quite similar, indicating that the data is not overfitted.
    

# Train Score
print(f'score on testing set: {lr.score(X_train,y_train)}')

# Test Score
print(f'score on training set: {lr.score(X_test,y_test)}')

# Cross Validation Score
cross_val_score(lr, X_train, y_train, cv=5).mean()

- Cross Validation and Test Score is quite simiilar in this case. 
- This indicates that its is a representative dataset.

# Scale our data.
# Relabeling scaled data as "Z" is common.
sc = StandardScaler()
Z_train = sc.fit_transform(X_train) # standard normal distribution
Z_test = sc.transform(X_test)

## Lasso Model

# Plot Lasso Model

lasso_model = Lasso(alpha =10)
# Fit.
lasso_model.fit(Z_train, y_train)
# Evaluate model using rmse.
y_pred_train = lasso_model.predict(Z_train)
y_pred_test = lasso_model.predict(Z_test)

rmse_train_l = mean_squared_error(y_true =y_train, y_pred=y_pred_train, squared=False)
rmse_test_l = mean_squared_error(y_true =y_test, y_pred=y_pred_test, squared=False)

# if mse, squared=True

print(rmse_train_l, rmse_test_l)

## Ridge Model

# Instantiate.
ridge_model = Ridge(alpha =10)
# Fit.
ridge_model.fit(Z_train, y_train)
# Evaluate model using rmse.
y_pred_train = ridge_model.predict(Z_train)
y_pred_test = ridge_model.predict(Z_test)

rmse_train_r = mean_squared_error(y_true =y_train, y_pred=y_pred_train, squared=False)
rmse_test_r = mean_squared_error(y_true =y_test, y_pred=y_pred_test, squared=False)

# if mse, squared=True

print(rmse_train_r, rmse_test_r)

# GridsearchCV Hyperparameter Tuning

## Ridge Regression

### Hyperparameter Tuning

# Instantiate Standard Scaler
ss = StandardScaler()

# Instantiate Model ridge for gridsearch cv
model_ridge = Ridge()


# Instantiate pipe
pipe_gs = Pipeline([
        ('ss', ss),
        ('rg', model_ridge)
        ])


# Instantiate Parameters
pipe_gs_param = {'ss__with_mean': [True, False], 
                 'ss__with_std': [True, False],    
                'rg__alpha':[1,5,10,15,20]                               
                }

# fit pipe
pipe_gs.fit(Z_train, y_train)

# pipe score
pipe_gs.score(X_test, y_test)

# define the grid search
Ridge_reg= GridSearchCV(pipe_gs, pipe_gs_param,cv=5,verbose=1)

# fit grid search mode
Ridge_reg.fit(Z_train,y_train)

Ridge_reg.best_score_

Ridge_reg.best_params_

### Modeling with shortlisted Hyperparameter for Ridge

# Instantiate Standard Scaler with new value
ss_ridge_gridcv = StandardScaler(with_mean = True, with_std = False)

# Standardise Predictors using ss with updated parameters
Z_train_ridge_gridcv = ss_ridge_gridcv.fit_transform(X_train) # standard normal distribution
Z_test_ridg_gridcv = ss_ridge_gridcv.transform(X_test)

# # Instantiate.
ridge_model_grid_cv = Ridge(alpha =1)
# Fit
ridge_model_grid_cv.fit(Z_train_ridge_gridcv, y_train)
# Evaluate model using rmse.
y_pred_train_ridge_grid_cv = ridge_model.predict(Z_train_ridge_gridcv)
y_pred_test_ridge_grid_cv = ridge_model.predict(Z_test_ridg_gridcv)

rmse_train_r = mean_squared_error(y_true =y_train, y_pred=y_pred_train_ridge_grid_cv, squared=False)
rmse_test_r = mean_squared_error(y_true =y_test, y_pred=y_pred_test_ridge_grid_cv, squared=False)
# if mse, squared=True

print(rmse_train_r, rmse_test_r)

- Observed that setting the Standard Scaler with the given parameters resulted in a higher rmse score.
- Next: rerun model where alpha is 1.

# Instantiate.
ridge_model = Ridge(alpha =1)
# Fit.
ridge_model.fit(Z_train, y_train)
# Evaluate model using rmse.
y_pred_train = ridge_model.predict(Z_train)
y_pred_test = ridge_model.predict(Z_test)

rmse_train_r = mean_squared_error(y_true =y_train, y_pred=y_pred_train, squared=False)
rmse_test_r = mean_squared_error(y_true =y_test, y_pred=y_pred_test, squared=False)

# if mse, squared=True

print(rmse_train_r, rmse_test_r)

- Observed that by setting rmse score to 1, the resulted rmse score is slightly better at 62485.527.

# Lasso Regression

### Hyper Parameter Tuning

lasso_alpha = [1,5,10,15,20]   

for i in lasso_alpha:   

    # Plot Lasso Model

    lasso_model = Lasso(alpha =i)
    # Fit.
    lasso_model.fit(Z_train, y_train)
    # Evaluate model using rmse.
    y_pred_train = lasso_model.predict(Z_train)
    y_pred_test = lasso_model.predict(Z_test)

    rmse_train_l = mean_squared_error(y_true =y_train, y_pred=y_pred_train, squared=False)
    rmse_test_l = mean_squared_error(y_true =y_test, y_pred=y_pred_test, squared=False)

    # if mse, squared=True
    print("alpha: ", i)
    print(rmse_train_l, rmse_test_l)

- It is observed that the RMSE score remains relatively high at around 62488.602 and 62322.953 for RMSE train and RMSE test data for Lasso Regression where alpha is 1.
- Next: Further analysis will be done on the features to better understand how different features affect RMSE.
    - Features will also be analysed to understand the trends.
